+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Each version section may have have subsections for: _Added_, _Changed_, _Removed
 ## [1.18.68]
 ### Added
 - Source factors can now be added to the embeddings instead of concatenated with `--source-factors-combine sum` (default: concat)
+- Fixed training crashes with `--learning-rate-decay-optimizer-states-reset initial` option.
 
 ## [1.18.67]
 ### Added