+++ CHANGELOG.md
@@ -10,8 +10,11 @@ Note that Sockeye has checks in place to not translate with an old model that wa
 
 Each version section may have have subsections for: _Added_, _Changed_, _Removed_, _Deprecated_, and _Fixed_.
 
+## [1.18.69]
+### Added
+- Source factors can now be added to the embeddings instead of concatenated with `--source-factors-combine sum` (default: concat)
+
 ## [1.18.68]
-### Fixed
 - Fixed training crashes with `--learning-rate-decay-optimizer-states-reset initial` option.
 
 ## [1.18.67]